Repair yrt formula
Likely due to a website update, the code used to determine a url for the
Viva system map PDF would fail to locate the anchor element. This commit
updates the code to match the changed website. Luckily, it only required
us to shift the index when obtaining an element from querySelectorAll.